use jwt extended

+++ resources/item.py
@@ -1,5 +1,6 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+# from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from models.item import ItemModel
 
 
@@ -37,7 +38,11 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -62,6 +67,15 @@ class Item(Resource):
 
 class ItemList(Resource):
 
+    @jwt_required(optional=True)
     def get(self):
+        user_id = get_jwt()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+        return {
+            'items': [item['name'] for item in items],
+            'message': 'More data available if you log in'
+        }
         # return {'items': [item.json() for item in ItemModel.query.all()]}
-        return {'items': list(map(lambda x: x.json(), ItemModel.query.all()))}
+        # return {'items': list(map(lambda x: x.json(), ItemModel.query.all()))}
